Added layer splitting, type decoding, relative/absolute etc
Lots of new things, I apologize for the formatting, somehow my editor
doesn't seem to grok this tabulation, or lack of.

The decoder now splits the object into 1) layers and 2) segment types.
Segment types are sorted by extrusion rate (speed) and whether they
are extruding or not.

materials and geometry are made for each of the layer, and all added to
the resulting object.

There is also support for G92, G90 and G91.

The decoder also computes the object min/max on the fly, excluding
non-extruding movements.

With this code, we can implement partial viewing of layers, and change the
alpha value and/or color relative to movement speed.

Signed-off-by: Michel Pollet <buserror@gmail.com>